FIR/UAST: commonize nullability annotation
